// src/pages/PDFJS/canvasTool/enum.ts
export enum DrawingTypeEnum {
  PEN = 'pen',
  TEXT = 'text',
  RECT = 'rect',
  ARC = 'arc',
  ARROWTRIANGLE = 'arrowTriangle',
  LINE = 'line',
  ARROWRADIUS = 'arrowRadius',
  ARROWDOUBLE = 'arrowDouble',
  CLOUD = 'cloud',
  ANCHOR = 'anchor',
}

export enum SelectTypeEnum {
  MULTI = 'multi',
  LASSO = 'lasso',
  CLUSTERSELECT = 'clusterSelect',
}

export enum CrossTypeEnum {
  BOTH = 'both',
  CONTENT = 'content',
  LINE = 'line',
}

export enum BtnEnum {
  DETAIL = 'detail',
  EDIT = 'edit',
  DELETE = 'delete',
}